Add module archive.mailarchive.
